+++ src/App.tsx
@@ -1,22 +1,28 @@
-import { useState } from "react";
 import Footer from "./components/Footer";
 import PrimaryNav from "./components/PrimaryNav";
 import PostsWrap from "./components/PostsWrap";
 import { initiatePostHog } from "./lib/analytics";
 import { BrowserRouter } from "react-router-dom";
 import { Toaster } from "react-hot-toast";
+import { SWRConfig } from "swr";
 import RepoSubmission from "./components/RepoSubmission";
 import GradBackground from "./components/GradBackground";
 import useSupabaseAuth from "./hooks/useSupabaseAuth";
 import Hero from "./components/Hero";
+import apiFetcher from "./hooks/useSWR";
 
 const App = (): JSX.Element => {
-  initiatePostHog();
   const { user } = useSupabaseAuth();
-  const [textToSearch] = useState("");
+
+  initiatePostHog();
 
   return (
-    <>
+    <SWRConfig
+      value={{
+        revalidateOnFocus: false,
+        fetcher: apiFetcher,
+      }}
+    >
       <Toaster position="top-right" />
 
       <BrowserRouter>
@@ -29,12 +35,12 @@ const App = (): JSX.Element => {
             <Hero />
           </GradBackground>
 
-          <PostsWrap textToSearch={textToSearch} />
+          <PostsWrap />
 
           <Footer />
         </div>
       </BrowserRouter>
-    </>
+    </SWRConfig>
   );
 };