+++ src/components/GridDisplay.tsx
@@ -5,7 +5,7 @@ export declare interface GridDisplayProps {
   activeLink: string | null;
   limit: number;
   handleLoadingMore: () => void;
-  fetchedData: DbRecomendation[];
+  fetchedData: DbRepo[];
   user?: User;
 }