+++ src/components/StackedAvatar.tsx
@@ -1,4 +1,5 @@
 import Avatar from "./Avatar";
+
 export declare interface StackedAvatarProps {
   contributors: DbContribution[];
 }